feat(ui): add note list screen for phone
