+++ src/screens/main.tsx
@@ -1,16 +1,9 @@
-import { Container } from '@/atoms'
-import HeaderBar from '@/components/header-bar'
-import MoveNoteSheet from '@/components/move-note-sheet'
-import NoteList from '@/components/note-list'
-import NoteListHeaderTitleBar from '@/components/note-list-header-title-bar'
-import useStickyHeader from '@/hooks/use-sticky-header'
 import { HomeDrawerParamList, RootStackParamList } from '@/navs'
-import { editingNoteIdAtom } from '@/states/editor'
 import { DrawerScreenProps } from '@react-navigation/drawer'
 import { CompositeScreenProps } from '@react-navigation/native'
 import { NativeStackScreenProps } from '@react-navigation/native-stack'
-import { useAtom } from 'jotai'
-import React, { useCallback, useRef, useState } from 'react'
+import React from 'react'
+import NoteListScreenForPhone from './note-list-phone'
 
 type Props = CompositeScreenProps<
   DrawerScreenProps<HomeDrawerParamList, 'Main'>,
@@ -18,56 +11,5 @@ type Props = CompositeScreenProps<
 >
 
 export default function MainScreen({ navigation }: Props) {
-  const refMoveNoteSheet = useRef<MoveNoteSheet>(null)
-  const {
-    handleNoteListLayout,
-    handleScroll,
-    headerBarStyle,
-    headerBarHeight
-  } = useStickyHeader()
-  const [concealNoteListItem, setConcealNoteListItem] = useState<
-    (() => void) | null
-  >(null)
-  const [, setEditingNoteId] = useAtom(editingNoteIdAtom)
-  const handleSidebarToggle = useCallback(() => {
-    navigation.toggleDrawer()
-  }, [navigation])
-  const handleNoteListItemPress = useCallback((noteId: string) => {
-    setEditingNoteId(noteId)
-    navigation.navigate('Detail')
-  }, [])
-  const handleNoteListItemSwipeLeft = useCallback(
-    (_noteId: string, conceal: () => void) => {
-      const { current: menu } = refMoveNoteSheet
-      if (menu) {
-        menu.show()
-        setConcealNoteListItem(() => conceal)
-      }
-    },
-    []
-  )
-  const handleMoveNoteSheetClose = useCallback(() => {
-    concealNoteListItem && concealNoteListItem()
-    setConcealNoteListItem(null)
-  }, [concealNoteListItem])
-  return (
-    <Container justifyContent="center" alignItems="center">
-      <NoteList
-        contentInsetTop={headerBarHeight}
-        onScroll={handleScroll}
-        onItemPress={handleNoteListItemPress}
-        onItemSwipeLeft={handleNoteListItemSwipeLeft}
-        ListHeaderComponent={NoteListHeaderTitleBar}
-      />
-      <HeaderBar
-        style={headerBarStyle}
-        onLayout={handleNoteListLayout}
-        onSidebarToggle={handleSidebarToggle}
-      />
-      <MoveNoteSheet
-        ref={refMoveNoteSheet}
-        onClose={handleMoveNoteSheetClose}
-      />
-    </Container>
-  )
+  return <NoteListScreenForPhone navigation={navigation} />
 }
